backend/requirements.txt
Django>=5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
django-filter>=23.5
mysqlclient>=2.2
python-decouple>=3.8
gunicorn>=22.0